requirements.txt
newspaper3k
urllib3==1.24.2
python-dateutil==2.8.0
requests==2.21.0
lxml==4.3.3
numpy==1.16.3
pandas==0.24.2
beautifulsoup4==4.7.1
mongoengine=0.18.2
alabaster==0.7.12
apturl==0.5.2
argcomplete==1.8.1
asn1crypto==0.24.0
astroid==2.1.0
attrs==19.1.0
autobahn==19.2.1
Automat==0.7.0
Babel==2.6.0
backcall==0.1.0
beautifulsoup4==4.6.0
billiard==3.6.0.0
bleach==3.1.0
blinker==1.4
Brlapi==0.6.6
bs4==0.0.1
cachetools==3.1.0
certifi==2018.11.29
cffi==1.12.0
chardet==3.0.4
checkbox-support==0.22
chrome-gnome-shell==0.0.0
Click==7.0
cloudpickle==0.8.0
coala-utils==0.5.1
command-not-found==0.3
confluent-kafka==0.11.6
constantly==15.1.0
cryptography==2.5
cssselect==1.0.3
cupshelpers==1.0
cx-Oracle==7.1.2
cycler==0.10.0
decorator==4.3.2
defer==1.0.6
defusedxml==0.5.0
distro-info===0.18ubuntu0.18.04.1
docutils==0.14
dt==0.2
entrypoints==0.3
expressvpn-python==1.0
fake-useragent==0.1.11
feedfinder2==0.0.4
feedparser==5.2.1
findspark==1.3.0
Flask==1.0.2
future==0.17.1
get==2018.11.19
google-api-python-client==1.7.8
google-auth==1.6.2
google-auth-httplib2==0.0.3
guacamole==0.9.2
html5lib==0.999999999
httplib2==0.12.1
hyperlink==19.0.0
idna==2.8
imagesize==1.1.0
incremental==17.5.0
ipykernel==4.10.0
ipyparallel==6.2.3
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
isort==4.3.4
itsdangerous==1.1.0
jedi==0.13.2
jeepney==0.4
jieba3k==0.35.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kazam==1.4.5
keyring==18.0.0
keyrings.alt==3.0
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.3.1
loginform==1.2.0
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
marshmallow==2.19.5
marshmallow-mongoengine==0.9.1
matplotlib==3.0.2
mccabe==0.6.1
mistune==0.8.4
mongoengine==0.18.2
mtranslate==1.6
nbconvert==5.4.1
nbformat==4.4.0
netifaces==0.10.4
newspaper3k==0.2.8
nltk==3.4.1
nose==1.3.7
notebook==5.2.2
numpy==1.16.1
numpydoc==0.8.0
oauth==1.0.1
oauthlib==3.0.1
olefile==0.45.1
onboard==1.4.1
packaging==19.0
padme==1.1.1
pandas==0.24.1
pandocfilters==1.4.2
parsel==1.5.1
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.1.0
plainbox==0.25
post==2018.11.20
progressbar==2.3
prompt-toolkit==2.0.8
protobuf==3.0.0
psutil==5.5.0
ptyprocess==0.6.0
public==2018.11.20
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycairo==1.16.2
pycodestyle==2.5.0
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
PyDispatcher==2.0.5
pyflakes==2.1.0
Pygments==2.3.1
pygobject==3.26.1
PyHamcrest==1.9.0
PyJWT==1.5.3
pylint==2.2.2
pymacaroons==0.13.0
pymongo==3.7.2
PyNaCl==1.1.2
pyOpenSSL==19.0.0
pyparsing==2.3.1
PyQt5==5.12
PyQt5-sip==4.19.14
pyRFC3339==1.0
PySocks==1.7.0
python-apt==1.6.4
python-dateutil==2.8.0
python-debian==0.1.32
python-gnupg==0.4.1
pytz==2018.9
pyxdg==0.25
PyYAML==3.12
pyzmq==17.1.2
QtAwesome==0.5.6
qtconsole==4.4.3
QtPy==1.6.0
query-string==2018.11.20
queuelib==1.5.0
reportlab==3.4.0
request==2018.11.20
requests==2.21.0
requests-file==1.4.3
requests-oauthlib==1.2.0
requests-unixsocket==0.1.5
rope==0.12.0
rsa==4.0
scikit-learn==0.20.2
scipy==1.2.1
scour==0.36
Scrapy==1.6.0
screen-resolution-extra==0.0.0
seaborn==0.9.0
SecretStorage==3.1.1
selenium==3.141.0
service-identity==18.1.0
simplegeneric==0.8.1
simplejson==3.13.2
six==1.12.0
slugify==0.0.1
snowballstemmer==1.2.1
Sphinx==1.8.4
sphinxcontrib-websupport==1.1.0
spyder==3.3.3
spyder-kernels==0.4.2
ssh-import-id==5.7
system-service==0.3
systemd-python==234
terminado==0.7
testpath==0.4.2
tinysegmenter==0.3
tldextract==2.2.1
tornado==4.5.3
traitlets==4.3.2
tweepy==3.7.0
Twisted==19.2.0
twitterscraper==1.2.0
txaio==18.8.1
typed-ast==1.3.1
ubuntu-drivers-common==0.0.0
Ubuntu-Make==18.5
ufw==0.36
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
unity-tweak-tool==0.0.7
uritemplate==3.0.0
urllib3==1.24.1
usb-creator==0.3.3
vboxapi==1.0
virtualenv==16.4.0
vpnotebook==0.1.3
vpython==7.4.7
w3lib==1.20.0
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==2.0.1
wikipedia==1.4.0
Wikipedia-API==0.5.1
wrapt==1.11.1
WTForms==2.2.1
wurlitzer==1.0.2
xdiagnose==3.8.8
xkit==0.0.0
xlrd==1.2.0
XlsxWriter==0.9.6
zope.interface==4.6.0
Flask==1.0.2
flask-restful==0.3.7
flask-marshmallow
